tests/submission/PhilipTaylor/canvas/toDataURL.zerosize.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: toDataURL.zerosize</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.toDataURL.html">toDataURL</a>.zerosize</h1>
<p class="desc">toDataURL on zero-size canvas returns 'data:,'</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.toDataURL.zerosize">toDataURL.zerosize</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="0" id="c" width="0"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("toDataURL on zero-size canvas returns 'data:,'");
_addTest(function(canvas, ctx) {

var data = canvas.toDataURL();
_assertSame(data, 'data:,', "data", "'data:,'");


});
</script>

<div id="log"></div>
</body></html>
